refactor(browser-eval): extract IIFE helpers to browser-eval-helpers.ts + fix CHANGELOG header
- Extract canParseAsExpression / isAlreadyWrappedIife / prepareBrowserEvalExpression
  into src/tools/browser-eval-helpers.ts (no side-effectful imports)
- browser.ts imports from new helpers file; shadow copies removed
- tests/unit/browser-eval-iife.test.ts now imports directly from helpers,
  eliminating shadow-copy maintenance risk
- Restore missing '## [0.14.0]' header in CHANGELOG.md
