+++ src/tools/browser.ts
@@ -28,6 +28,7 @@ import { narrateParam } from "./_narration.js";
 import type { RichBlock } from "../engine/uia-diff.js";
 import { evaluatePreToolGuards, buildEnvelopeFor } from "../engine/perception/registry.js";
 import { runActionGuard, isAutoGuardEnabled, validateAndPrepareFix, consumeFix } from "./_action-guard.js";
+import { canParseAsExpression, isAlreadyWrappedIife, prepareBrowserEvalExpression } from "./browser-eval-helpers.js";
 
 // ─────────────────────────────────────────────────────────────────────────────
 // Schemas
@@ -826,57 +827,6 @@ function safeStringifyEval(value: unknown): string {
 function safeCloneForTransport(value: unknown): unknown {
   try { return JSON.parse(safeStringifyEval(value)); }
   catch { return "[Unserializable]"; }
-}
-
-type AsyncFunctionConstructor = new (...args: string[]) => (...args: unknown[]) => Promise<unknown>;
-const AsyncFunction = Object.getPrototypeOf(async function () { /* constructor only */ }).constructor as AsyncFunctionConstructor;
-
-function isAlreadyWrappedIife(expression: string): boolean {
-  const normalized = expression
-    .trim()
-    .replace(/^;/, "")
-    .replace(/;+\s*$/, "");
-
-  if (!/^\s*\(\s*(?:async\s+function\b|function\b|async\s*\([^)]*\)\s*=>|\([^)]*\)\s*=>)/.test(normalized)) {
-    return false;
-  }
-
-  return canParseAsExpression(normalized);
-}
-
-function canParseAsExpression(expression: string): boolean {
-  try {
-    new AsyncFunction(`return (\n${expression}\n);`);
-    return true;
-  } catch {
-    return false;
-  }
-}
-
-function prepareBrowserEvalExpression(expression: string): string {
-  if (isAlreadyWrappedIife(expression)) return expression;
-
-  if (canParseAsExpression(expression)) {
-    return `;(async () => (\n${expression}\n))()`;
-  }
-
-  const serializedExpression = JSON.stringify(expression);
-  return `;(async () => {
-  const __mcpExpression = ${serializedExpression};
-  try {
-    return eval(__mcpExpression);
-  } catch (__mcpEvalError) {
-    if (
-      __mcpEvalError instanceof SyntaxError ||
-      (__mcpEvalError instanceof Error && __mcpEvalError.name === "EvalError")
-    ) {
-      return (async () => {
-${expression}
-      })();
-    }
-    throw __mcpEvalError;
-  }
-})()`;
 }
 
 export const browserEvalHandler = async ({
